Fix granules that show inline videos.
